pyproject.toml
[project]
name = "pi_sensors"
version = "0.1.0"
description = "Raspberry Pi sensor dashboard — FastAPI backend"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    # Web framework
    "fastapi>=0.115.0",
    "uvicorn>=0.32.0",
    "loguru>=0.7.3",
    "pydantic-settings>=2.6.0",
    # Sensors
    "adafruit-circuitpython-bme680>=3.7.14",
    "sparkfun-qwiic-pir>=0.0.4",
    "smbus2>=0.5.0",
    # Audio
    "sounddevice>=0.5.1",
    "numpy>=2.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/pi_sensors"]

[dependency-groups]
dev = [
    "pytest>=8.0.0",
    "ruff>=0.9.0",
    "pytest-mock>=3.14.0",
    "httpx>=0.28.0",    # required by Starlette TestClient
]

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.ruff]
target-version = "py313"
line-length = 120
fix = true

[tool.ruff.lint]
select = [
    "YTT", "S", "B", "A", "C4", "T10", "SIM",
    "I", "C90", "E", "W", "F", "PGH", "UP", "RUF", "TRY",
]
ignore = [
    "E501", "E731", "B008", "TRY301",
]

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["S101"]

[tool.ruff.format]
preview = true
indent-style = "space"
quote-style = "double"
